rename to flare
