+++ flare/flare/examples.py
